Fix steering vector calculation
Steering vectors depend on the frequency of the signal-of-interest, but
not on the central frequency of a DAQ as it is now. So lets compensate
for the discrepancy. Also, for convenience lets set VFOs frequencies to
the central one on receiver parameters update. Finally, we should use
radius of UCA for steering vectors calculation, but since
fe91116aa7b26f42bd92ca6e10675eef376f4336 we use inter element spacing
instead.