Add option to use automatic gain control
It would be highly advantageous to have any form of automatic gain
control, since signal properties can hardly be predicted and,
generally, could be time-dependent. Since Kraken's ADCs have rather
low dynamic range, detecting signals below its noise level and
maximizing utilization of the ADCs' dynamic range requires
manual adjustments of the gains by the operator and so, complicates
workflow to large extent.

So lets go ahead and enable RF (LNA and mixer) hardware-based
automatic gain control (AGC). Eventhough we have a librtlsdr fork that
allows for Variable Gain Amplifier (VGA) AGC driven by the RTL2832U
feedback [1]. I was not successful to enable it (is it wired up?). As
expected, pure RF AGC can overshoot and overdrive ADC. So some tuning
needs to be done to corresponding voltage threshould levels. I believe,
enabling VGA AGC is also essential here, because not only it provides
+40 dB of gain, but can also attenuate IF signal by up-to -12 dB to
prevent ADC overdrive. Little is known about properties of VGA AGC
feedback loop, though, and it might not be suited for short narrowband
signals at the end of the day.

Last, but not least, since AGCs are working independently in each
channel, it is theoretically possible for different amounts of gain to
be set in each channel. This might lead to incorrect DoA estimation. So
users are warned about this if `Auto` gain is selected.

[1] https://codingspirit.de/librtlsdr-driver.pdf