Unconditionally save settings after initialization
For clients that alter `settings.json` to control the app, it is
important to have all the available settings listed in the file. This
becomes an issue when migrating to software versions that introduce
new settings, yet re-using old `settings.json` file. To mitigate this
problem, lets update `settings.json` file right after initialization to
make sure it contains all the settings available in the corresponding
software version.