Hide `Reset Configuration` button in `System Control` panel
It seems like current users tend to press `Reset Configuration` because
they assume it is former `Save Configuration` that used to be place in
exactly the same place. This might significantly break the workflow in
some mission-critical scenarious. So lets safeguard this function by
moving it to the (hidden) `System Control` panel. Also, lets output
`doa.xml` irrespective of the output format setting, but retain
dummy `DF-Aggregator` option to not confuse current and prospective
users.